+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 + Add `--skip-attr` used to exclude migration between two nodes that have same attributes
 + Add `--max-shard-size` to skip shards larger than a certain size
 + Add `--node-role` to filter nodes by their role
++ add `--max-recovery-per-node` to avoid moving shards to a node that is envolved in recovery
 + Enhanced the skip-attr check, to avoid swap of a shard that can have a replica in a node that have same attribute (used for some costrain).