+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.7
+# 1.0
 
 + Add `--use-shard-id` to use shard id for algorithm rebalance instead index name
 + Add `--skip-attr` used to exclude migration between two nodes that have same attributes